Fix Int overflow in Effect.Fog default color
0x80FFFFFF exceeds Int.MAX_VALUE as a literal and failed to compile
under compileDebugKotlin; needs the same .toInt() conversion already
used for Effect.Particles' default color.

Verified end-to-end: ./gradlew assembleDebug now builds a debug APK
and testDebugUnitTest passes, now that dl.google.com/maven.google.com
are reachable and the Android SDK (platform 34, build-tools 34.0.0)
is installed.

+++ app/src/main/java/com/audiovisualizer/render/effects/Effect.kt
@@ -21,7 +21,7 @@ sealed class Effect {
     data class Fog(
         val density: Float = 0.5f,
         val zone: SpawnZone = SpawnZone.FullScreen,
-        val color: Int = 0x80FFFFFF
+        val color: Int = 0x80FFFFFF.toInt()
     ) : Effect()
 
     data class Glow(
